refactor: move Recipe Loaders to loaders/recipe
