+++ src/main/kotlin/com/github/trcdevelopers/clayium/common/loaders/recipe/LatheRecipeLoader.kt
@@ -1,12 +1,10 @@
-package com.github.trcdevelopers.clayium.common.recipe.loader
+package com.github.trcdevelopers.clayium.common.loaders.recipe
 
 import com.github.trcdevelopers.clayium.common.clayenergy.ClayEnergy
 import com.github.trcdevelopers.clayium.common.items.metaitem.MetaItemClayParts
 import com.github.trcdevelopers.clayium.common.recipe.registry.CRecipes
-import com.github.trcdevelopers.clayium.common.unification.material.CMaterials
 import com.github.trcdevelopers.clayium.common.unification.material.CMaterials.clay
 import com.github.trcdevelopers.clayium.common.unification.material.CMaterials.denseClay
-import com.github.trcdevelopers.clayium.common.unification.ore.OrePrefix
 import com.github.trcdevelopers.clayium.common.unification.ore.OrePrefix.Companion.block
 import com.github.trcdevelopers.clayium.common.unification.ore.OrePrefix.Companion.cylinder
 import com.github.trcdevelopers.clayium.common.unification.ore.OrePrefix.Companion.disc
