+++ src/main/kotlin/com/github/trcdevelopers/clayium/common/loaders/recipe/MetaTileEntityLoader.kt
@@ -1,0 +1,4 @@
+package com.github.trcdevelopers.clayium.common.loaders.recipe
+
+object MetaTileEntityLoader {
+}
