+++ src/main/kotlin/com/github/trcdevelopers/clayium/common/loaders/recipe/SolarClayFabricatorRecipeLoader.kt
@@ -1,4 +1,4 @@
-package com.github.trcdevelopers.clayium.common.recipe.loader
+package com.github.trcdevelopers.clayium.common.loaders.recipe
 
 import com.github.trcdevelopers.clayium.common.recipe.registry.CRecipes
 import com.github.trcdevelopers.clayium.common.unification.material.CMaterials
