+++ src/main/kotlin/com/github/trcdevelopers/clayium/common/loaders/recipe/WireDrawingRecipeLoader.kt
@@ -1,11 +1,11 @@
-package com.github.trcdevelopers.clayium.common.recipe.loader
+package com.github.trcdevelopers.clayium.common.loaders.recipe
 
 import com.github.trcdevelopers.clayium.common.clayenergy.ClayEnergy
-import com.github.trcdevelopers.clayium.common.items.metaitem.MetaItemClayParts
 import com.github.trcdevelopers.clayium.common.recipe.registry.CRecipes
 import com.github.trcdevelopers.clayium.common.unification.material.CMaterials
 import com.github.trcdevelopers.clayium.common.unification.ore.OrePrefix
 import net.minecraft.init.Items
+import kotlin.collections.withIndex
 
 object WireDrawingRecipeLoader {
     fun registerRecipes() {
