--- src/main/kotlin/com/github/trcdevelopers/clayium/common/recipe/loader/MetaTileEntityLoader.kt
@@ -1,4 +1,0 @@
-package com.github.trcdevelopers.clayium.common.recipe.loader
-
-object MetaTileEntityLoader {
-}
